refactor(cli): move image counter to common
