+++ tosho/src/impl/kmkc/download.rs
@@ -8,6 +8,7 @@ use tosho_kmkc::{
     KMClient,
 };
 
+use crate::r#impl::common::check_downloaded_image_count;
 use crate::term::Terminal;
 use crate::{
     cli::ExitCode,
@@ -40,30 +41,6 @@ pub(crate) struct KMDownloadCliConfig {
 
     pub(crate) no_ticket: bool,
     pub(crate) no_point: bool,
-}
-
-fn check_downloaded_image_count(image_dir: &PathBuf, extension: &str) -> Option<usize> {
-    // check if dir exist
-    if !image_dir.exists() {
-        return None;
-    }
-
-    // check if dir is dir
-    if !image_dir.is_dir() {
-        return None;
-    }
-
-    // check how many .avif files in the dir
-    let mut count = 0;
-    for entry in std::fs::read_dir(image_dir).unwrap() {
-        let entry = entry.unwrap();
-        let path = entry.path();
-        if path.is_file() && path.extension().unwrap() == extension {
-            count += 1;
-        }
-    }
-
-    Some(count)
 }
 
 fn create_chapters_info(title: &TitleNode, chapters: Vec<EpisodeNode>) -> MangaDetailDump {
