+++ tosho/src/impl/rbean/download.rs
@@ -15,6 +15,7 @@ use crate::{
     cli::ExitCode,
     r#impl::{
         clean_filename,
+        common::check_downloaded_image_count,
         models::{ChapterDetailDump, MangaDetailDump},
     },
     term::{ConsoleChoice, Terminal},
@@ -68,30 +69,6 @@ pub(crate) struct RBDownloadConfigCli {
 
     /// Parallel download
     pub(crate) parallel: bool,
-}
-
-fn check_downloaded_image_count(image_dir: &PathBuf, extension: &str) -> Option<usize> {
-    // check if dir exist
-    if !image_dir.exists() {
-        return None;
-    }
-
-    // check if dir is dir
-    if !image_dir.is_dir() {
-        return None;
-    }
-
-    // check how many .avif files in the dir
-    let mut count = 0;
-    for entry in std::fs::read_dir(image_dir).unwrap() {
-        let entry = entry.unwrap();
-        let path = entry.path();
-        if path.is_file() && path.extension().unwrap() == extension {
-            count += 1;
-        }
-    }
-
-    Some(count)
 }
 
 fn create_chapters_info(title: &Manga, chapters: Vec<Chapter>) -> MangaDetailDump {
