+++ tosho/src/impl/sjv/download.rs
@@ -7,6 +7,7 @@ use tosho_sjv::{
     SJClient, SJPlatform,
 };
 
+use crate::r#impl::common::check_downloaded_image_count;
 use crate::{
     cli::ExitCode,
     r#impl::{
@@ -36,30 +37,6 @@ pub(crate) struct SJDownloadCliConfig {
     ///
     /// Used only when `no_input` is `true`.
     pub(crate) end_at: Option<u32>,
-}
-
-fn check_downloaded_image_count(image_dir: &PathBuf, extension: &str) -> Option<usize> {
-    // check if dir exist
-    if !image_dir.exists() {
-        return None;
-    }
-
-    // check if dir is dir
-    if !image_dir.is_dir() {
-        return None;
-    }
-
-    // check how many .avif files in the dir
-    let mut count = 0;
-    for entry in std::fs::read_dir(image_dir).unwrap() {
-        let entry = entry.unwrap();
-        let path = entry.path();
-        if path.is_file() && path.extension().unwrap() == extension {
-            count += 1;
-        }
-    }
-
-    Some(count)
 }
 
 fn create_chapters_info(title: &MangaDetail, chapters: Vec<MangaChapterDetail>) -> MangaDetailDump {
